WIP: Removed propagator in favor or handling in chainedPropagator
